+++ BackEnd/booking_service/src/main/java/iuh/fit/booking_service/client/refund/RefundClient.java
@@ -1,0 +1,15 @@
+/*
+ * @ (#) RèundClient.java 1.0 4/26/2025
+ *
+ * Copyright (c) 2025 IUH.All rights reserved
+ */
+package iuh.fit.booking_service.client.refund;
+
+/*
+ * @description
+ * @author : Nguyen Truong An
+ * @date : 4/26/2025
+ * @version 1.0
+ */
+public interface RefundClient {
+}
