+++ FrontEnd/src/components/admin/booking/bookingFilters.js
@@ -1,0 +1,61 @@
+import React from 'react';
+import { BOOKING_STATUSES, PAYMENT_FILTERS } from './constants';
+
+const BookingFilters = ({ searchTerm, setSearchTerm, statusFilter, setStatusFilter, paymentFilter, setPaymentFilter }) => {
+  return (
+    <div className="mb-4 flex flex-col md:flex-row gap-4">
+      <div className="flex-1">
+        <label htmlFor="search" className="sr-only">
+          Search
+        </label>
+        <div className="relative rounded-md shadow-sm">
+          <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+            <svg className="h-5 w-5 text-gray-400" fill="currentColor" viewBox="0 0 20 20">
+              <path
+                fillRule="evenodd"
+                d="M8 4a4 4 0 100 8 4 4 0 000-8zM2 8a6 6 0 1110.89 3.476l4.817 4.817a1 1 0 01-1.414 1.414l-4.816-4.816A6 6 0 012 8z"
+                clipRule="evenodd"
+              />
+            </svg>
+          </div>
+          <input
+            type="text"
+            id="search"
+            className="focus:ring-indigo-500 focus:border-indigo-500 block w-full pl-10 sm:text-sm border-gray-300 rounded-md p-2 border"
+            placeholder="Search by name, email, or booking code"
+            value={searchTerm}
+            onChange={(e) => setSearchTerm(e.target.value)}
+          />
+        </div>
+      </div>
+      <div className="w-full md:w-auto">
+        <select
+          className="block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm p-2 border"
+          value={statusFilter}
+          onChange={(e) => setStatusFilter(e.target.value)}
+        >
+          {BOOKING_STATUSES.map((status) => (
+            <option key={status} value={status}>
+              {status === 'ALL' ? 'All Statuses' : status}
+            </option>
+          ))}
+        </select>
+      </div>
+      <div className="w-full md:w-auto">
+        <select
+          className="block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm p-2 border"
+          value={paymentFilter}
+          onChange={(e) => setPaymentFilter(e.target.value)}
+        >
+          {PAYMENT_FILTERS.map((filter) => (
+            <option key={filter} value={filter}>
+              {filter === 'ALL' ? 'All Payments' : filter === 'PENDING' ? 'Pending' : 'Others'}
+            </option>
+          ))}
+        </select>
+      </div>
+    </div>
+  );
+};
+
+export default BookingFilters;
