+++ FrontEnd/src/components/admin/booking/bookingList.js
@@ -1,0 +1,120 @@
+import React from 'react';
+import StatusBadge from './statusBadge';
+import PaymentStatus from './paymentStatus';
+import { formatPrice } from './utils';
+
+const BookingList = ({ bookings, isLoading, handleBookingClick, setSelectedBooking, setIsCancelModalOpen, setNewStatus, setCancelReason, CANCEL_REASONS }) => {
+  return (
+    <div className="overflow-x-auto">
+      <table className="min-w-full divide-y divide-gray-200">
+        <thead className="bg-gray-50">
+          <tr>
+            {[
+              'Booking Code',
+              'Customer',
+              'Tour Code',
+              'Status',
+              'Booking Date',
+              'Payment Due',
+              'Total Price',
+              'Actions',
+            ].map((header) => (
+              <th
+                key={header}
+                className="px-6 py-3 text-left text-xs font-medium text-white uppercase tracking-wider"
+              >
+                {header}
+              </th>
+            ))}
+          </tr>
+        </thead>
+        <tbody className="bg-white divide-y divide-gray-200">
+          {isLoading ? (
+            <tr>
+              <td colSpan={8} className="px-6 py-4 text-center text-sm text-gray-500">
+                <svg
+                  className="animate-spin mx-auto h-8 w-8 text-indigo-600"
+                  xmlns="http://www.w3.org/2000/svg"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                >
+                  <circle
+                    className="opacity-25"
+                    cx="12"
+                    cy="12"
+                    r="10"
+                    stroke="currentColor"
+                    strokeWidth="4"
+                  />
+                  <path
+                    className="opacity-75"
+                    fill="currentColor"
+                    d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+                  />
+                </svg>
+              </td>
+            </tr>
+          ) : bookings.length === 0 ? (
+            <tr>
+              <td colSpan={8} className="px-6 py-4 text-center text-sm text-gray-500">
+                No bookings found. Try adjusting your search or filter criteria.
+              </td>
+            </tr>
+          ) : (
+            bookings.map((booking) => (
+              <tr
+                key={booking.id}
+                onClick={() => handleBookingClick(booking)}
+                className="cursor-pointer hover:bg-gray-50"
+              >
+                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                  {booking.bookingCode || 'N/A'}
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap">
+                  <div className="text-sm font-medium text-gray-900">
+                    {booking.user?.fullName || 'N/A'}
+                  </div>
+                  <div className="text-sm text-gray-500">{booking.user?.email || 'N/A'}</div>
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                  {booking.tour?.tourCode || 'N/A'}
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap">
+                  <StatusBadge status={booking.bookingStatus} />
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                  {booking.bookingDate ? new Date(booking.bookingDate).toLocaleString() : 'N/A'}
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                  <PaymentStatus booking={booking} />
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                  {formatPrice(booking.totalPrice)}
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                  {booking.bookingStatus !== 'CANCELLED' &&
+                    booking.bookingStatus !== 'COMPLETED' && (
+                      <button
+                        onClick={(e) => {
+                          e.stopPropagation();
+                          setSelectedBooking(booking);
+                          setNewStatus('CANCELLED');
+                          setCancelReason(CANCEL_REASONS[0].value);
+                          setIsCancelModalOpen(true);
+                        }}
+                        className="text-red-600 hover:text-red-800"
+                      >
+                        Cancel
+                      </button>
+                    )}
+                </td>
+              </tr>
+            ))
+          )}
+        </tbody>
+      </table>
+    </div>
+  );
+};
+
+export default BookingList;
