+++ FrontEnd/src/pages/admin/dashboard.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import AdminNavbar from '../../components/admin/adminNavbar';
-import BookingTable from '../../components/admin/bookingTable';
+import BookingTable from '../../components/admin/booking/bookingTable';
 import TourTable from '../../components/admin/tourTable';
 import UserTable from '../../components/admin/userTable';
 import SettingModal from '../../components/admin/settingModal';
@@ -29,29 +29,25 @@ function Dashboard() {
   };
 
   return (
-    <div className="flex flex-col h-screen">
-      {/* Toast Container */}
+    <div className="flex flex-col h-screen bg-gray-100 ">
       <ToastContainer position="top-right" autoClose={3000} hideProgressBar={false} />
-
-      {/* Navbar */}
       <AdminNavbar isSidebarOpen={isSidebarOpen} setIsSidebarOpen={setIsSidebarOpen} />
-
-      <div className="flex flex-1">
-        {/* Sidebar */}
+      <div className="flex flex-1 overflow-hidden">
         <AdminSidebar
           activeTab={activeTab}
           setActiveTab={setActiveTab}
           openSetting={() => setIsSettingOpen(true)}
           isSidebarOpen={isSidebarOpen}
+          setIsSidebarOpen={setIsSidebarOpen}
         />
-
-        {/* Main Content */}
-        <main className="flex-1 p-4 sm:ml-64 mt-16 overflow-auto">
+        <main
+          className={`flex-1 p-4 mt-16 transition-all duration-300 ${
+            isSidebarOpen ? 'ml-64' : 'ml-0'
+          } overflow-auto`}
+        >
           {renderTable()}
         </main>
       </div>
-
-      {/* Setting Modal */}
       {isSettingOpen && <SettingModal onClose={() => setIsSettingOpen(false)} />}
     </div>
   );
